requirements.txt
django-crispy-forms
djangorestframework
djangorestframework-bulk
